Added Methods And Revised Coding Style
Revised coding style in all the classes .In this case added a blank line
to the end of each class's java file in files that lacked such a line,
and checked indentation.

In class com.sol.util.ArrayUtils changed the name of the methods
#toPrimitiveWrapper to the name #toBoxedArray to increase readability,
added corresponding parallel methods of #toPrimitiveArray and
#toBoxedArray since not in all cases converting an array to/from
primitive in parallel is desired.

For readability reasons also removed a number of private methods from
com.sol.util.ArrayUtils (they contained one line and it's doubtful they
actually increased readability.)