Added Class ArrayHandler And Fixed Copyrights
Added the class ArrayHandler. Class is meant to be a substitute to the
java.lang.reflect.Array class. The latter use JNI methods that, for
performance reasons are better implemented in java.

Fixed The copyrights of said class from Eclipse copyrights to Apache
Licanse 2.0-like copyrights, since the latter is the license of the
project.