Change class name and enhanced safety of certain classes
Changed the name of ArrayUtils to ArrayUtilities since the ApacheCommons
library contains a class by the name ArrayUtils (the goal of the change
is to avoid import mix-up of the classes).

In the classes SoftReferenceContainer and WeakReferenceContainer added a
check for null references in the constructor.

In class ReferenceContainer removed assertion from constructor since a
null value is already checked.

In Class ActionBot removed a superfluous pair of parentheses from the
expiciteWait method.